requirements.txt
openai
requests
pymupdf
arxiv
numpy
sentence-transformers
pyyaml
beautifulsoup4
selenium
webdriver-manager
Pillow
